test(client): test SearchTracksScreen
